requirements.txt
langchain
langchain-openai
faiss-cpu
sentence-transformers
pydantic
python-dotenv
tiktoken
fastapi
uvicorn
numpy
pandas
scikit-learn
